add redux for task id
move code from SimulationProperties to SimulationScreen2
